+++ src/main/scala/com/decisionbrain/cplex/cp/NumVar.scala
@@ -78,6 +78,27 @@ class NumVar(v: IloNumVar)(implicit model: CpModel) extends NumExpr(v) with Adda
     * @return the CPLEX numeric variable
     */
   override def getIloAddable(): IloAddable = v
+
+  /**
+    * Return a character string that represents the integer variable.
+    *
+    * @return a character string
+    */
+  override def toString() : String = {
+    val strBuilder = new StringBuilder()
+    if (getName().isDefined) strBuilder.append(getName().get)
+    strBuilder.append("[")
+    val vmin = getLB()
+    val vmax = getUB()
+    strBuilder.append(vmin)
+    if (vmin < vmax) {
+      strBuilder.append("..")
+      strBuilder.append(vmax)
+    }
+    strBuilder.append("]")
+    strBuilder.toString()
+  }
+
 }
 
 object NumVar {