+++ src/test/scala/com/decisionbrain/cplex/cp/LearningCurveTest.scala
@@ -11,19 +11,19 @@ import org.scalatest.junit.JUnitRunner
 import org.scalatest.{FunSuite, Matchers}
 
 @RunWith(classOf[JUnitRunner])
-class SchedSetupTest extends FunSuite with Matchers {
+class LearningCurveTest extends FunSuite with Matchers {
 
   val epsilon = 1e-6
 
-  test("SchedSetup") {
-    val model = SchedSetup.build()
+  test("LearningCurve") {
+    val model = LearningCurve.build()
 
-    val status = SchedSetup.solve()
+    val status = LearningCurve.solve()
 
     status should equal(true)
     // Note: a solution with objective 174 can be obtained by increasing the fail limit (see method solve)
-    model.getObjectiveValue() should be <= 199.0
+    model.getObjectiveValue() should be <= 269.0
 
     model.end()
-}
+  }
 }